Added switches: start fuel and liters
Changes have been made to update various user interactions and enhance the conversion feature in the code. Several variables have been promoted to instance level, along with the addition of a new conversion factor variable to aid in conversion. Visible properties of certain widgets are also toggled on or off based on the value of the "starting fuel switch". A new method, toggle_conversion_factor, has been introduced to adjust the conversion factor based on whether liters or gallons are selected.