Refactor HelloWorld app by replacing unused components
Two libraries, 'httpx' and 'numpy', were removed as they were not used in this version of the app. The input and 'say hello' button have been replaced with a slider for adjusting starting fuel, which is hidden until toggled by the user. The new slider is connected to an updated method for calculating delta volume.